backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv

chromadb

crewai

httpx
